+++ src/gui/Point.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstdint>
 #include <utility>
 
 namespace gui {
@@ -12,13 +11,13 @@ struct Point {
           : X(0),
             Y(0) {
     }
-    constexpr Point(uint16_t x,
-          uint16_t y)
+    constexpr Point(size_t x,
+          size_t y)
           : X(x),
             Y(y) {
     }
-    uint16_t X;
-    uint16_t Y;
+    size_t X;
+    size_t Y;
 };
 
 constexpr Point operator+(const Point& lhs, const Point& rhs) {
